Extract and re-use formatUses()
