+++ src/useCases/mcaWritingStyleFormatter.ts
@@ -1,7 +1,21 @@
+import { IUse } from "../entities/IUse";
+
 export const formatDate = (dateString: string): string => {
   const date = new Date(dateString);
   const [, month, day, year] = date.toDateString().split(" ");
   return `${parseInt(day)} ${month} ${year.slice(2)}`;
+};
+
+export const formatUses = (uses: IUse[]): string =>
+  uses.reduce((formattedUses, use, index, uses) => {
+    if (index === uses.length - 1) return formattedUses + formatUse(use);
+    return formattedUses + formatUse(use) + ", ";
+  }, "");
+
+const formatUse = (use: IUse): string => {
+  const formattedActivity = titleCase(use.activity);
+  const formattedPurpose = use.purpose ? ` (${titleCase(use.purpose)})` : "";
+  return formattedActivity + formattedPurpose;
 };
 
 export const titleCase = (text: string): string => {
